ch_14/requirements.txt

beautifulsoup4==4.9.1
jsonschema==3.2.0
httpx==0.18.1
pytest-httpx==0.12.0
pyyaml==5.3.1
pillow==8.0.1
